Refactor workload cluster out of cluster
